requirements.txt
Flask==2.3.3
APScheduler==3.10.4
openpyxl==3.1.2
python-dotenv==1.0.0
Flask-WTF==1.1.1
flask-talisman==1.1.0
flask-limiter==3.5.0
gunicorn==21.2.0
gevent==23.9.1
email-validator==2.1.0.post1
requests==2.31.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
pytz==2023.3.post1
tzlocal==5.0.1
